+++ CHANGELOG.md
@@ -5,8 +5,8 @@ dd mmm yyyy
 - Add: API endpoint for apps that returns all characters from a player account.
 - Add: API endpoint for apps that accepts a corporation ID and returns a list of all player IDs that have a 
   character in the corporation.
+- Member tracking: added filter options for member list
 - Small improvements for UI, frontend and documentation.
-
 
 ## 1.2.1